filters.py
moved the EntryInstance filter into its own file for import into the
views.py file
